Introduce FrontendGraphPrinter
