+++ tools/src/tensil/tools/compiler/EmitContext.scala
@@ -6,5 +6,6 @@ package tensil.tools.compiler
 case class EmitContext(
     hir: HIR,
     mm: MemoryManager,
-    outputNames: Seq[String]
+    outputNames: Seq[String],
+    graphPrinter: Option[FrontendGraphPrinter] = None
 )
