+++ tools/src/tensil/tools/compiler/Frontend.scala
@@ -3,7 +3,6 @@
 
 package tensil.tools.compiler
 
-import tensil.tools.GraphPrinter
 import tensil.tools.data.Shape
 
 abstract class Frontend {
@@ -11,6 +10,4 @@ abstract class Frontend {
   def rewrite(program: Seq[String]): Seq[Emitter]
 
   def mkConstsDimensions(shape: Shape): MemoryDimensions
-
-  val graphPrinter: Option[GraphPrinter]
 }
